move photos' position calcultating into a separate file
